Reorganised spawning module
