+++ src/behaviour/mod.rs
@@ -11,7 +11,6 @@ pub mod lifecycle;
 pub mod metabolism;
 pub mod movement;
 pub mod sight;
-pub mod spawning;
 pub mod thinking;
 pub mod timers;
 
@@ -30,7 +29,6 @@ pub fn time_step_system_set() -> SystemSet {
         .with_system(growth::process_growers_system.after(thinking::thinking_system))
         .with_system(lifecycle::transition_to_adult_system)
         .with_system(lifecycle::transition_to_hatching_system)
-        .with_system(lifecycle::spawn_plant_system)
         .with_system(eating::attempted_to_eat_system.after(eating::process_eaters_system))
         .with_system(growth::attempted_to_grow_system.after(growth::process_growers_system))
         .with_system(growth::existence_system)
@@ -48,7 +46,6 @@ pub fn slow_behaviour_system_set() -> SystemSet {
         .with_system(eating::eating_system.after(eating::attempted_to_eat_system))
         .with_system(lifecycle::lay_egg_system.after(lifecycle::process_layers_system))
         .with_system(growth::grow_bug_system.after(growth::attempted_to_grow_system))
-        .with_system(spawning::nearest_spawner_system)
 }
 
 pub fn metabolism_system_set() -> SystemSet {
