+++ src/lib.rs
@@ -1,4 +1,7 @@
-use bevy::prelude::{App, CoreStage, Plugin, SystemStage};
+use bevy::{
+    prelude::{App, CoreStage, Plugin, SystemSet, SystemStage},
+    time::FixedTimestep,
+};
 
 mod attributes;
 mod behaviour;
@@ -7,7 +10,20 @@ mod config;
 mod ecosystem;
 mod mind;
 mod setup;
+mod spawning;
 mod ui;
+
+pub fn slow_system_set() -> SystemSet {
+    SystemSet::new()
+        .with_run_criteria(FixedTimestep::step(0.1))
+        .with_system(spawning::nearest_spawner_system)
+}
+
+pub fn time_step_system_set() -> SystemSet {
+    SystemSet::new()
+        .with_run_criteria(FixedTimestep::step(config::TIME_STEP as f64))
+        .with_system(spawning::spawn_plant_system)
+}
 
 pub struct GenesisPlugin;
 
@@ -18,8 +34,7 @@ impl Plugin for GenesisPlugin {
 
         let config_instance = config::WorldConfig::global();
 
-        let spawners =
-            behaviour::spawning::Spawners::from_configs(&config_instance.spawners).unwrap();
+        let spawners = spawning::Spawners::from_configs(&config_instance.spawners).unwrap();
 
         app.add_stage_after(CoreStage::Update, CLEAN_UP, SystemStage::parallel())
             .insert_resource(config::BACKGROUND)
@@ -38,11 +53,13 @@ impl Plugin for GenesisPlugin {
             .add_system_set(behaviour::metabolism_system_set())
             .add_system_set(ui::global_statistics_system_set())
             .add_system_set(ui::regular_saver_system_set())
+            .add_system_set(slow_system_set())
+            .add_system_set(time_step_system_set())
             .add_system_to_stage(CoreStage::Last, ui::save_on_close)
             .add_system_set_to_stage(CLEAN_UP, behaviour::despawn_system_set())
             .insert_resource(ecosystem::Ecosystem::new(config_instance.world_energy))
             .insert_resource(spawners)
-            .insert_resource(behaviour::lifecycle::PlantSizeRandomiser::new(
+            .insert_resource(spawning::PlantSizeRandomiser::new(
                 config_instance.plant_size_range,
             ));
     }
